Separate `executeMarry` and `executeDivorce`
