+++ client/src/commands/social/MarryCommand.ts
@@ -1,31 +1,9 @@
 import { bot } from "../../index";
-import ComponentInteractionContext from "../../structures/commands/ComponentInteractionContext";
 import { createCommand } from "../../structures/commands/createCommand";
 import { createButton, createCustomId, createActionRow } from "../../utils/discord/Component";
 import { ApplicationCommandOptionTypes, ButtonStyles } from "discordeno/types";
 import { User } from "discordeno/transformers";
-
-const executeMarry = async (ctx: ComponentInteractionContext) => {
-    const userData = await bot.database.getUser(ctx.author.id);
-    const partnerData = await bot.database.getUser(ctx.interaction.user.id);
-
-    userData.marriedWith = ctx.user.id;
-    userData.marriedDate = new Date();
-    partnerData.marriedWith = ctx.author.id;
-    partnerData.marriedDate = new Date();
-    await userData.save();
-    await partnerData.save();
-
-    ctx.foxyReply({
-        content: ctx.makeReply("❤", bot.locale("commands:marry.accepted")),
-        components: [createActionRow([createButton({
-            customId: createCustomId(0, ctx.interaction.data.targetId, ctx.commandId),
-            label: bot.locale("commands:marry.accept"),
-            style: ButtonStyles.Success,
-            disabled: true
-        })])],
-    })
-}
+import executeMarry from "../../structures/commands/modules/executeMarry";
 
 const MarryCommand = createCommand({
     path: '',
